Reorganize scripts/, moving some useful functions to script_utils.js, and some common consts to script_consts.js
