+++ scripts/script_consts.js
@@ -1,0 +1,10 @@
+const path = require('path');
+
+const build_dir = path.resolve(__dirname, '../build');
+
+const pdf_name = 'stephen-oneil-resume.pdf';
+
+module.exports = {
+  pdf_name,
+  build_dir
+};
